requirements.txt
langchain>=0.2.12
langgraph>=0.2.14
pydantic>=2.8.2
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
httpx>=0.27.0
networkx>=3.3
numpy>=1.26.4
pandas>=2.2.2
faiss-cpu>=1.8.0
tiktoken>=0.7.0
openai>=1.41.1
